tijdelijke push
push voordat pc crashed

+++ Test periode 2/Assets/Scripts/Ro/Enemy/Alien.cs
@@ -7,7 +7,7 @@ using UnityEngine.AI;
 public class Alien : MonoBehaviour
 {
     public GameObject player;
-    public int ahealth;
+    public float ahealth;
     public int damage;
     public float maxFollowRange, minFollowRange;
     public float attackRange;
